requirements.txt
uvicorn==0.32.0
fastapi==0.115.3
asyncpg==0.30.0
alembic==1.13.3
loguru==0.7.2
PyJWT==2.8.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
SQLAlchemy~=2.0.36
pydantic~=2.9.2
fastapi-users[sqlalchemy]
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-dotenv==0.5.2
httpx==0.28.1